[24] T5: extract channel.go, final cleanup, update() one-line dispatch
Create channel.go with errChannelClosed, channelSubscribeCmd,
channelReadNextCmd, and handleChannelEvent.

Extract remaining >5-line inline handlers from update() into named methods:
handleTick (session.go), handleLabelCheckResult/handleLabelsEnsured/
handleIssuesLoaded/handleIssueConfirmed (tracker_ops.go),
handleChannelEvent (channel.go).

Move handleFocusSwitch/handleLoopSlotsKey to launch.go (slot operations).
Move RunServerInit/serverInitCmds to session.go (session scan init).
Move tickCmd and session-only constants to session.go.

Final: model.go 2433 → 860 lines. All 19 tests pass, go vet clean.
